requirements.txt
# MikrotikAPI-BF v3.0 — Python 3.8+ requirements
# Install: pip install -r requirements.txt

# ── Core HTTP / networking ─────────────────────────────────────────────────
requests>=2.28.0
urllib3>=1.26.0

# ── Terminal UX ────────────────────────────────────────────────────────────
colorama>=0.4.6

# ── SSH post-login validation ─────────────────────────────────────────────
paramiko>=2.7.0

# ── SOCKS proxy support ────────────────────────────────────────────────────
PySocks>=1.7.1

# ── YAML configuration ────────────────────────────────────────────────────
PyYAML>=6.0

# ── HTML parsing (fingerprinting) ─────────────────────────────────────────
beautifulsoup4>=4.12.0
